requirements.txt
djangorestframework
django-ratelimit==3.0.0
pyjwt==1.7.1
river
redis
dill
